django_taskmaster_backend/backend/requirements.txt
Django>=5.1,<7.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.4.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
